requirements.txt
ollama>=0.3.0
chromadb>=0.5.0
python-docx>=1.1.0
